Version 2.0.5.8
- Refactoring

+++ library/src/main/java/com/github/cris16228/library/http/image_loader/MemoryCache.java
@@ -82,6 +82,6 @@ public class MemoryCache {
     private long sizeInBytes(Bitmap bitmap) {
         if (bitmap == null)
             return 0;
-        return (long) bitmap.getByteCount();
+        return bitmap.getByteCount();
     }
 }
